--- app/src/test/java/ProfileDao.kt
@@ -1,10 +1,0 @@
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-interface ProfileDao {
-    @Insert
-    suspend fun insert(profile: Profile)
-
-}
