--- app/src/test/java/TaskDao.kt
@@ -1,12 +1,0 @@
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-interface TaskDao {
-    @Insert
-    suspend fun insert(task: Task)
-
-    @Query("SELECT * FROM tasks")
-    suspend fun getAllTasks(): List<Task>
-}
