--- app/src/test/java/UserDao.kt
@@ -1,12 +1,0 @@
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-interface UserDao {
-    @Insert
-    suspend fun insert(user: User)
-
-    @Query("SELECT * FROM users")
-    suspend fun getAllUsers(): List<User>
-}
